+++ QueryStringTest.php
@@ -17,9 +17,13 @@ use League\Uri\Exceptions\SyntaxError;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
+use stdClass;
 use Stringable;
+use TypeError;
+use ValueError;
 
 use function date_create;
+use function tmpfile;
 
 use const PHP_QUERY_RFC1738;
 use const PHP_QUERY_RFC3986;
@@ -516,5 +520,216 @@ final class QueryStringTest extends TestCase
             'query' => 'key[]9[]=bar',
             'expected' => 'key%5B0%5D=bar',
         ];
+    }
+
+    /**
+     * @param non-empty-string $separator
+     */
+    #[DataProvider('providesVariablesToCompose')]
+    public function test_it_can_compose_a_query_string(object|array $variable, string $separator, int $encoding, ?string $expected): void
+    {
+        self::assertSame($expected, QueryString::compose($variable, $separator, $encoding));
+    }
+
+    public static function providesVariablesToCompose(): iterable
+    {
+        yield 'null if the variable is empty' => [
+            'variable' => [],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => null,
+        ];
+
+        yield 'null if the object properties are not accessible' => [
+            'variable' => new stdClass(),
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => null,
+        ];
+
+        yield 'basic encoding from php-src tests 1' => [
+            'variable' =>  ['foo' => 'bar', 'baz' => 1, 'test' => "a ' \" ", 'abc', 'float' => 10.42, 'true' => true, 'false' => false],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => 'foo=bar&baz=1&test=a+%27+%22+&0=abc&float=10.42&true=1&false=0',
+        ];
+
+        yield 'basic encoding from php-src tests 2 - with a different separator' => [
+            'variable' =>  ['foo' => 'bar', 'baz' => 1, 'test' => "a ' \" ", 'abc', 'float' => 10.42, 'true' => true, 'false' => false],
+            'separator' => ';',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => 'foo=bar;baz=1;test=a+%27+%22+;0=abc;float=10.42;true=1;false=0',
+        ];
+
+        $data = new class () implements Stringable {
+            public string $public = 'input';
+            protected string $protected = 'hello';
+            private string $private = 'world';
+            public function __toString(): string
+            {
+                return $this->private;
+            }
+        };
+
+        yield 'basic encoding from php-src tests 3 - with object' => [
+            'variable' =>  $data,
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => 'public=input',
+        ];
+
+        yield 'basic encoding from php-src tests 3 - with null object' => [
+            'variable' =>  new stdClass(),
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => null,
+        ];
+
+        $data = new class () implements Stringable {
+            public function __toString(): string
+            {
+                return 'Stringable';
+            }
+        };
+
+        yield 'basic encoding from php-src tests 4 - with object with is just stringable' => [
+            'variable' =>  ['hello', $data],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => '0=hello',
+        ];
+
+        yield 'basic encoding from php-src tests 5 - with object with is just stringable' => [
+            'variable' =>  $data,
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => null,
+        ];
+
+        $o = new class () {
+            public mixed $public = 'input';
+        };
+        $nested = clone $o;
+        $o->public = $nested;
+
+        yield 'basic encoding from php-src tests 6 - nested object' => [
+            'variable' =>  $o,
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => 'public%5Bpublic%5D=input',
+        ];
+
+        $obj = new stdClass();
+        $obj->name = 'homepage';
+        $obj->page = 1;
+        $obj->sort = 'desc,name';
+
+        yield 'basic encoding from php-src tests 7 - stdClass' => [
+            'variable' =>  $obj,
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => 'name=homepage&page=1&sort=desc%2Cname',
+        ];
+
+        yield 'basic encoding from php-src tests 8 - array' => [
+            'variable' =>  [
+                20,
+                5 => 13,
+                '9' => [
+                    1 => 'val1',
+                    3 => 'val2',
+                    'string' => 'string',
+                ],
+                'name' => 'homepage',
+                'page' => 10,
+                'sort' => [
+                    'desc',
+                    'admin' => [
+                        'admin1',
+                        'admin2' => [
+                            'who' => 'admin2',
+                            2 => 'test',
+                        ],
+                    ],
+                ],
+            ],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => '0=20&5=13&9%5B1%5D=val1&9%5B3%5D=val2&9%5Bstring%5D=string&name=homepage&page=10&sort%5B0%5D=desc&sort%5Badmin%5D%5B0%5D=admin1&sort%5Badmin%5D%5Badmin2%5D%5Bwho%5D=admin2&sort%5Badmin%5D%5Badmin2%5D%5B2%5D=test',
+        ];
+
+        yield 'basic encoding from php-src tests 8 - array with rfc1738 encoding' => [
+            'variable' =>  [
+                'name' => 'main page',
+                'sort' => 'desc,admin',
+                'equation' => '10 + 10 - 5',
+            ],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC1738,
+            'expected' => 'name=main+page&sort=desc%2Cadmin&equation=10+%2B+10+-+5',
+        ];
+
+        yield 'basic encoding from php-src tests 8 - array with rfc3986 encoding' => [
+            'variable' =>  [
+                'name' => 'main page',
+                'sort' => 'desc,admin',
+                'equation' => '10 + 10 - 5',
+            ],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC3986,
+            'expected' => 'name=main%20page&sort=desc%2Cadmin&equation=10%20%2B%2010%20-%205',
+        ];
+
+        yield 'basic encoding from php-src tests 8 - with null' => [
+            'variable' =>  [null],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC3986,
+            'expected' => '0',
+        ];
+
+        $v = 'value';
+        $ref = &$v;
+
+        yield 'basic encoding from php-src tests 8 - with reference' => [
+            'variable' =>  [$ref],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC3986,
+            'expected' => '0=value',
+        ];
+
+        yield 'bug resolution in php-src tests 9 - float conversion' => [
+            'variable' =>  ['x' => 1E+14, 'y' => '1E+14'],
+            'separator' => '&',
+            'encoding' => PHP_QUERY_RFC3986,
+            'expected' => 'x=1.0E%2B14&y=1E%2B14',
+        ];
+    }
+
+    public function test_it_throws_if_a_object_recursion_is_detected(): void
+    {
+        $recursive = new class () {
+            public mixed $public = 'input';
+        };
+
+        $recursive->public = $recursive;
+        $this->expectException(ValueError::class);
+
+        QueryString::compose($recursive);
+    }
+
+    public function test_it_throws_if_a_array_recursion_is_detected(): void
+    {
+        $recursive = [];
+        $recursive['self'] = &$recursive;
+        $this->expectException(ValueError::class);
+
+        QueryString::compose($recursive);
+    }
+
+    public function test_it_throws_if_a_resource_is_present(): void
+    {
+        $this->expectException(TypeError::class);
+
+        QueryString::compose([tmpfile()]);
     }
 }